requirements.txt
# ============================================================
# Research Assistant AI — Dependencies
# ============================================================
# Install: pip install -r requirements.txt
# ============================================================

# --- Google Gemini API (LLM + Embeddings) ---
# google-generativeai is deprecated — use google-genai instead
google-genai>=1.68.0

# --- Data Validation ---
pydantic>=2.7.1

# --- Web UI ---
streamlit>=1.35.0

# --- Data Handling ---
pandas>=2.2.2

# --- Environment Variables ---
python-dotenv>=1.0.1

# --- LangChain (Document Loading + Vector Store) ---
langchain>=0.2.5
langchain-community>=0.2.5
langchain-text-splitters>=0.2.2
langchain-core>=0.2.9

# --- Vector Similarity Search ---
faiss-cpu>=1.8.0

# --- PDF Parsing (used by LangChain PyPDFLoader) ---
pypdf>=4.2.0